rename: SessionManager 클래스를 session 패키지 아래로 이동
