+++ tomcat/src/main/java/org/apache/coyote/http11/handler/GetHttp11MethodHandler.java
@@ -9,10 +9,10 @@ import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
-import org.apache.catalina.SessionManager;
 import org.apache.coyote.header.HttpCookie;
 import org.apache.coyote.header.HttpMethod;
 import org.apache.coyote.session.Session;
+import org.apache.coyote.session.SessionManager;
 import org.apache.coyote.util.RequestExtractor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
