+++ tomcat/src/main/java/org/apache/coyote/http11/handler/PostHttp11MethodHandler.java
@@ -8,10 +8,10 @@ import static org.apache.coyote.header.HttpMethod.POST;
 import java.util.HashMap;
 import java.util.Map;
 import nextstep.jwp.model.User;
-import org.apache.catalina.SessionManager;
 import org.apache.coyote.header.HttpCookie;
 import org.apache.coyote.header.HttpMethod;
 import org.apache.coyote.session.Session;
+import org.apache.coyote.session.SessionManager;
 import org.apache.coyote.util.RequestExtractor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
